+++ NEWS.md
@@ -8,6 +8,7 @@ dispRity v1.9.3 (2025-03-06)
 ### MINOR IMPROVEMENTS
 
  * `chrono.subsets` now can recycle the `$tree` component of a `dispRity` object.
+ * The output of `multi.ace` using `estimation.details` with specific character arguments (e.g. `"detail_name"`) is now sorted in the order `output[[tree_ID]][[character_ID]]$detail_name`, rather than the previous version order that was `output[[tree_ID]]$detail_name[[character_ID]]`.
 
 ### BUG FIXES